refactor: outbound.search 패키지 내부 정리
