+++ api/src/main/java/vook/server/api/outbound/search/demo/DemoTermSearchParams.java
@@ -1,4 +1,4 @@
-package vook.server.api.outbound.search;
+package vook.server.api.outbound.search.demo;
 
 import com.meilisearch.sdk.SearchRequest;
 import lombok.Builder;
