+++ api/src/main/java/vook/server/api/web/routes/demo/DemoWebService.java
@@ -3,8 +3,8 @@ package vook.server.api.web.routes.demo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.outbound.search.DemoTermSearchResult;
-import vook.server.api.outbound.search.DemoTermSearchService;
+import vook.server.api.outbound.search.demo.DemoTermSearchResult;
+import vook.server.api.outbound.search.demo.DemoTermSearchService;
 import vook.server.api.web.routes.demo.reqres.SearchTermRequest;
 import vook.server.api.web.routes.demo.reqres.SearchTermResponse;
