+++ api/src/main/java/vook/server/api/web/routes/demo/reqres/SearchTermRequest.java
@@ -2,7 +2,7 @@ package vook.server.api.web.routes.demo.reqres;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
-import vook.server.api.outbound.search.DemoTermSearchParams;
+import vook.server.api.outbound.search.demo.DemoTermSearchParams;
 
 import java.util.List;
